+++ backend/prisma/migrations/20260725022407_cooperative_logo/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "Cooperative" ADD COLUMN     "logo" BYTEA,
+ADD COLUMN     "logoMimeType" TEXT;